Fixed DuckDB workaround and made it implicit Now columns from parquet files are correctly selected and the workaround is applied automatically.
